some of the technical material on polymorphism appears in the figures. in general, disregard the text in the document it is a left over that has not been edited.
the work here is incomplete.  Here are some notes from the blackboard from
a meeting with Yitzhak and Kathleen.  In some cases I have forgotten i
exactly what the comments mean or am not sure I agree:

-- base types should include \bottom (default value).  \bottom
may not be the exact perfect technical notion.  I am not sure.
I usually think of bottom as undefined/non-termination.  is that what we
want?

-- set constraints should test the underlying PD first to
avoid computing with \bottom.

-- add explicit roll/unroll coercions -- this will clarify
how the translation of recursive types work and make the definition
easier to read.

-- add explicit /\alpha for type abstraction to the translation everywhere.
also add corresponding type application e [t] everywhere.  Again,
this is for clarity and for consistency.

-- should we add explicit typing annotations on ordinary \lambda's?
this might be the right thing to do if we have explicit roll/unroll
and /\ alpha.

-- eliminate unnecessary sum values in the rep (Principle:  programmer
uses the parse descriptor to detect errors not the representation.)

-- delete normalization of recursion?  no longer necessary with the
new way of doing recursion with pd_hdr's being associated with
recursive type as well as the underlying type?

-- correlation relation is inductive on the representation.  (not
sure what this point means or if it means changing something...)

Dave