edited translation of recursive types;  I inserted explicit type annotations and explicit rolls to clarify the invariants.
I wasn't sure what the pd for a recursive type should be. I made it a copy
of the pd for the underlying type -- I hope that is correct.

I deleted the contractiveness annotations in the kinding rules.

Aside from type normalization and the error correctness theorem,
I think the rest is correct.