updated error correlation relation and normalization.
eliminated the mu a.t unrolling normalization rule -- I don't think
it is necessary now.

added mu a.t to the normal types.

added clause for mu a.t to the correlation relation.  The main
condition is the equality of the mu pd header with header of
the underlying type.

Dave