Added a draft of the related work section. It is relatively long -- approximately 1 full page.  I imagine there are places to cut here if necessary.
notice I used in the directory-local pads.bib for the bibliography.
I did this so that I can copy a make-able version of the paper home
and back just by copying the directory.  when we are done writing the paper,
if someone wants to copy the updated bibliography back to the "common"
area, that is fine.

Dave