This is the first pass for a tool to get data from urls (remote locations, files, etc). It is built using the multi interface to the libcurl library (http://curl.haxx.se/libcurl/). I haven't yet integrated this with pads.
The general idea is to have two threads: a requester and a getter. One thread (padsc) makes requests for urls as it encounters them (ie. during parsing). A second thread running in the background gets these requests and makes the results available as soon as possible.

get.c contains the implementation for the work queue and the interface to libcurl to do the actual work. The two main functions are:

void *RunGetData();            /* Loop that waits for requests and gets data */

void MakeRequest(char *url);   /* Make a request for data from url */


testget.c does a simple test by creating two threads. One calls RunGetData(), and the other makes calls to MakeRequest() intermingled with calls to usleep().

I have no idea how this works on various other platforms. Hopefully libcurl is pretty standard everywhere!