Implements a Performance Regression Test
The performance regression test is built on top of the existing regression
test with several caveats:

1. The measurement technology is too imprecise for almost all regression
tests.  (warnings are produced in this case)

2. Measurements are host-specific and therefore a baseline for the host on
which you test is required. BUT

3. Assistance is available for creating a baseline through the use of an
environment variable: PADS_PERF_REGRESS_BASELINE.

To assist with constructing the baseline database, the tests may be run
with the environment variable PADS_PERF_REGRESS_BASELINE set to:
- "AUTO" to have this script automatically create missing baseline files.
- "FORCE" to have this script overwrite baseline files.
- an empty string (or unset) for normal operation

4. If the test is deemed to have been measured with sufficient accuracy
then the measurement is compared to the baseline AND, if it exceeds a
threshold (set in rules.mk), then a "DIFFERS" message is triggered
(i.e. the performance test is deemed to have failed).