working on final journal version. edited introduction slightly.  I find the claims that we made earlier about having a "semantics framework for describing all ad hoc data processing languages" too broad.  We're describing type-based description languages like pads but not, say, giving a semantics to Perl or Python which certainly can and are used to process ad hoc data.
I have also been working on rearranging the related work.  Some of that was
phrased as a comparison between PADS and other tools, which I do not think
we want to do exactly.  The paper is (in my mind) not exactly a paper
just about PADS, rather it is a paper on how to do a semantics of
languages like PADS and also PACKETTYPES and DATASCRIPT, etc.  Just
direct comparisons of PADS vs. X did not quite fit into the context
of the rest of the paper (IMHO).  I tried to make it more comparisons between
"class" of languages ....

anyway, I am not quite done yet but I decided to check in a partial version...