edited section 6.5
I'm not sure if we should add the schedule analysis part or delete it.  Deleting
saves space.  Having it in to me seems like it's reaching (we're desperate for
an argument that we don't have) because we're using something we haven't done.
Also, it suggests to me that the combinator is perhaps just wrong -- the combinator
should instead just take an argument that is a schedule that is the schedule it forces its subcomponents to have so that there's just no way to define a badly scheduled pair of feeds...