Michel Blockelet, Summer intern at Princeton University in June-July 2008 developed the first implementation of PADS for Haskell under the supervision of David Walker.
Some of the key design ideas:

* use of built-in GADTs to represent the abstract syntax of PADS
   * interestingly, we considered many options for type-directed
     programming (many different ones exist in Haskell).  We settled
     on using GADTs, forgetting that that is what Yitzhak had done for
     PADS/ML -- that probably means that Yitzhak really made the right
     choice!  Though the implementation in Haskell is trivial compared with
     the trickiness that Yitzhak had to figure out to do the encoding.

   * though it probably looks easy now that we've done it (and perhaps
     it was! :-)  The difficult part was coming up with a solution that
     allowed (1) arbitrary data types to be representations (2) dependency
     including dependency on values with arbitrary types (3) annotations
     carrying values with arbitrary types (4) keeping things "well-typed"
     so that the type checker actually enforces useful invariants (eg:
     avoiding universal datatype definitions) (5) making it possible to
     use the representation to write all generic tools

   * all tools, including parsers and printers are coded easily using
     induction over the structure of the GADT

   * thanks to Stephanie Weirich for discussions about the various options
     we could use and help with GADTs in Haskell

* encoding of arbitrary data types using existential types and lists of
  constructor encodings where each constructor is encoded as a pair of
  injection and projection functions.  Turns out we reinvented an idea
  Stephanie Weirich uses in her RepLib library for generic programming.

* inclusion of "annotations" in the PADS descriptions that can be used to
customize the actions of particular tools.  These annotations carry values
of arbitrary types.  Moreover, each new tool can define its own new set of
annotations.  The implementation technique involves using the type class
Typeable to simulate an extensible datatype.  To define a new set of
annotations for a new tool, merely define a new datatype and derive the
Typeable type class.  In the tool, given an annotation, use type-safe
cast to determine if the annotation has a type associated with the tool.
If so, process the annotation however desired.  If not, discard the
annotation.  Thanks to the Haskell-cafe mailing lists for suggestions
on how to encode extensible datatypes using Typeable and existential
types (we figured out the existentials ourselves but weren't aware of
the existence of Typeable until we asked the Haskell gurus.

* we reimplemented regular expressions and added "tags" as suggested
by Ville Laurikari:

http://citeseer.ist.psu.edu/350334.html

* we implemented parser, printer, xml generator.  All of these tools
were more or less trivial to implement once we figured out the right
representation for the PADS abstract syntax.  There are things we
need to fix with the tools (eg: the xml pds are done differently
than the other tools; the parser and error recovery is a mess) but
this is "just engineering" :-).  Basically, the overall implementation
strategy and all the typing now "just works".  The features of Haskell
give a huge leg up on C (obviously!) but also it would appear on O'Caml.

Dave