formatting changes in evaluation.tex
don't use the [h] option in figures -- it puts them in the middle
of pages which takes more space (and doesn't look good).  If you actually
want something inline in the text, don't use a figure, just inline it
between paragraphs.

also cut the big 173-line description.  we can put that back in if a final version gives us more space.  It didn't give us back as much space as I'd hoped due to bad paragraph placement/figure placement throughout the paper.

revamped the explaination of the pads description in section 2.  going to
work on the intro more.