I believe I fixed the relation so that it now uses lookahead for arrays and lookahead is properly propagated elsewhere.
moreover, I constrained Pint and Regexp base types such that they are
required to have longest match semantics.

eg:

you can't match 12 and then have a number (eg: 3) at the beginning of the lookahead.
this will have the effect that we can cut down on the possible alternative parses
and if we didn't have fixed-width dependent strings in the input, we could use
lex like the current learning system does to do tokenization as a pre-pass.

Dave