Reworked the bidirectional semantics.
This is mostly to a stable state, but there is a problem with the
constraints. I think I need to be able to assume

  phi(F1++F2) implies phi(F1) and phi(F2)

to show that the dependent pair construct <x:s1,s2> is
well-behaved. However, the constraint produced by the
glob-comprehension specification

    (fun F ->
       ...
       F(r) = Directory{n1,...,nk'}
       ...)

does not obey it.

I'm out of brain cells for the night. Perhaps the solution will be
clear with fresh eyes in the morning.