Added Featherweight POSIX section.
Changed formalization to use advisory locks.

Changed the Makefile and main a little, adding pervasives.sty and kinda copying one of Nate's earlier papers. Mostly to later be able to steal his typesetting for the POSIX section. Feel free to undo and I'll put all of it in main instead.